requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
asyncio==3.4.3
asynctest==0.13.0
attrs==23.2.0
cached-property==1.5.2
certifi==2024.2.2
charset-normalizer==3.3.2
dataclasses-json==0.6.4
deepgram-sdk==3.0.1
distro==1.9.0
exceptiongroup==1.2.0
frozenlist==1.3.3
greenlet==3.0.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.6
importlib-metadata==6.7.0
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.11
langchain-community==0.0.25
langchain-core==0.1.29
langchain-text-splitters==0.0.1
langsmith==0.1.19
marshmallow==3.19.0
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.24.4
openai==1.13.3
orjson==3.9.15
packaging==23.2
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==0.21.1
python-wordpress-xmlrpc==2.3
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==2.0.28
tenacity==8.2.3
tiktoken==0.6.0
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.7
verboselogs==1.7
websockets==11.0.3
yarl==1.9.4
zipp==3.15.0
